Refactored profile code into components
Currently working on mobile graphic optimizations
