+++ app/_components/top_artists/displays/DisplayPFP.tsx
