+++ app/_components/top_artists/displays/TimeFrame.tsx
@@ -1,0 +1,38 @@
+"use client";
+9
+import { Dispatch, SetStateAction } from "react";
+import textColouring from "../text_colouring/TextColouring";
+import { LocalTrack, RegularTrack } from "@/app/_types/data";
+
+export default function Timeframe({range, setRange, setInformation}: 
+    {
+        range: "short_term" | "medium_term" | "long_term" | undefined, 
+        setRange: Dispatch<SetStateAction<"short_term" | "medium_term" | "long_term">>,
+        setInformation: Dispatch<SetStateAction<LocalTrack | RegularTrack | undefined>>;
+    }){
+
+    return(
+        <div className="text-center h-fit ">
+            <span className="text-3xl text-fontBlue font-semibold">
+            Adjust Timeframe
+            </span>
+            <div className="flex mt-4 w-full justify-around text-2xl">
+                {/* Setting information only to undefined if user is actually changing into a different time range, otherwise 
+                results in unnecessary updates in the DOM */}
+                <button className={"transition-all duration-300 "+textColouring("short_term", range)}
+                onClick={()=>{if(range != "short_term"){setInformation(undefined);} setRange("short_term");}}>
+                    Short Term
+                </button>
+                <button className={"transition-all duration-300 "+textColouring("medium_term", range)}
+                onClick={()=>{if(range != "medium_term"){setInformation(undefined);} setRange("medium_term");}}>
+                    Medium Term
+                </button>
+                <button className={"transition-all duration-300 "+textColouring("long_term", range)}
+                onClick={()=>{if(range != "long_term"){setInformation(undefined);} setRange("long_term");}}>
+                    Long Term
+                </button>
+            </div>
+        </div>
+    )
+
+}
