+++ app/_components/top_artists/text_colouring/TextColouring.ts
@@ -1,0 +1,3 @@
+export default function textColouring(name: string, selected?: string) {
+    return name == selected ? "text-purple" : "hover:scale-110";
+  }
