--- app/page.module.css
@@ -1,1 +1,0 @@
-/* Dont care abotu making it look nice */
